refactor/3 @Value -> @ConfigurationProperties으로 변경
