+++ src/main/java/kr/gravy/gravy/email/configuration/MailConfiguration.java
@@ -1,6 +1,7 @@
 package kr.gravy.gravy.email.configuration;
 
-import org.springframework.beans.factory.annotation.Value;
+import kr.gravy.gravy.configuration.properties.MailProperties;
+import lombok.RequiredArgsConstructor;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.mail.javamail.JavaMailSender;
@@ -9,34 +10,30 @@ import org.springframework.mail.javamail.JavaMailSenderImpl;
 import java.util.Properties;
 
 @Configuration
+@RequiredArgsConstructor
 public class MailConfiguration {
 
     public static final int POSSIBLE_REQUEST_TIME = 180;
 
-    @Value("${spring.mail.username}")
-    private String username;
-
-    @Value("${spring.mail.password}")
-    private String password;
+    private final MailProperties mailProperties;
 
     @Bean
     public JavaMailSender javaMailSender() {
         JavaMailSenderImpl mailSender = new JavaMailSenderImpl();
-        mailSender.setHost("smtp.gmail.com");
-        mailSender.setPort(587);
-
-        mailSender.setUsername(username);
-        mailSender.setPassword(password);
+        mailSender.setHost(mailProperties.host());
+        mailSender.setPort(mailProperties.port());
+        mailSender.setUsername(mailProperties.username());
+        mailSender.setPassword(mailProperties.password());
 
         Properties props = mailSender.getJavaMailProperties();
-        props.put("mail.transport.protocol", "smtp");
-        props.put("mail.smtp.auth", "true");
-        props.put("mail.smtp.starttls.enable", "true");
+        props.put("mail.transport.protocol", mailProperties.transport().protocol());
+        props.put("mail.smtp.auth", String.valueOf(mailProperties.smtp().auth()));
+        props.put("mail.smtp.starttls.enable", String.valueOf(mailProperties.smtp().starttls().enable()));
 
-        // 메일 서버 타임아웃 설정 (5초)
-        props.put("mail.smtp.connectiontimeout", "5000"); // 연결 타임아웃
-        props.put("mail.smtp.timeout", "5000");           // 읽기 타임아웃
-        props.put("mail.smtp.writetimeout", "5000");      // 쓰기 타임아웃
+        // Duration을 밀리초로 변환하여 JavaMail Properties에 적용
+        props.put("mail.smtp.connectiontimeout", String.valueOf(mailProperties.smtp().connectiontimeout().toMillis()));
+        props.put("mail.smtp.timeout", String.valueOf(mailProperties.smtp().timeout().toMillis()));
+        props.put("mail.smtp.writetimeout", String.valueOf(mailProperties.smtp().writetimeout().toMillis()));
 
         return mailSender;
     }
